+++ stage2/protocol/chainload.h
@@ -1,5 +1,9 @@
 #pragma once
 #include <stage2/driver/disk.h>
+#include <stdnoreturn.h>
 #include <stdbool.h>
+#include <stdint.h>
 
-bool chainload(disk_t disk);
+#define CHAINLOAD_SIGNATURE 0xAA55
+
+void chainload(disk_t disk);